Modified fields, getters and setters
