+++ src/main/java/pl/kowalczyk/maciej/java/app/bookingapp/model/Guest.java
@@ -66,6 +66,14 @@ public class Guest {
         this.name = name;
     }
 
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
     public List<Reservation> getReservations() {
         return this.reservations;
     }
